Desafio Mocks

+++ Entregable/src/controllers/productController.js
@@ -41,9 +41,14 @@ export const getProduct = async (req, res) => {
 
 export const createProduct = async (req, res)  => {
     try {
-        const prod = req.body
-        const message = await productModel.create(prod)
+        if (req.user.rol == "Admin"){
+            const prod = req.body
+            const message = await productModel.create(prod)
             res.status(201).send(message)
+        } else {
+            res.status(403).send("Usuario no autorizado")
+        }
+
     } catch(error) {
         res.status(500).send(`Error interno del servidor al crear producto: ${error}`);
     }
@@ -51,10 +56,14 @@ export const createProduct = async (req, res)  => {
 
 export const updateProduct = async (req, res) => {
     try {
-        const { id } = req.params.pid;
-        const prodUpdate = req.body;
-        const prod = await productModel.findByIdAndUpdate(id, prodUpdate)
-        res.status(200).send(prod)
+        if (req.user.rol == "Admin"){
+            const { id } = req.params.pid;
+            const prodUpdate = req.body;
+            const prod = await productModel.findByIdAndUpdate(id, prodUpdate)
+            res.status(200).send(prod)
+        } else {
+            res.status(403).send("Usuario no autorizado") 
+        }
     } catch(error) {
         res.status(500).send(`Error interno del servidor al actualizar producto: ${error}`);
     }
@@ -62,9 +71,14 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     try {
-        const { id } = req.params;
-        const message = await productModel.findByIdAndDelete(id)
-        res.status(200).send(message)
+        if (req.user.rol == "Admin"){
+            const { id } = req.params;
+            const message = await productModel.findByIdAndDelete(id)
+            res.status(200).send(message)
+        } else {
+            res.status(403).send("Usuario no autorizado")
+        }
+
         
     } catch(error) {
         res.status(500).send(`Error interno del servidor al eliminar producto: ${error}`);
